feat: 增加useCascader hooks，提取公共函数复用
