+++ src/cascader/index.ts
@@ -1,4 +1,5 @@
 export { cascaderProps, default as NCascader } from './src/Cascader'
 export type { CascaderProps, CascaderSlots } from './src/Cascader'
 export { default as NCascaderPanel } from './src/CascaderPanel'
+export type { CascaderPanelSlots } from './src/CascaderPanel'
 export type { CascaderInst, CascaderOption } from './src/interface'
